@@ -1,27 +1,27 @@
 <?php
 
-class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
+class ConceptMappingPropertyValueTest extends PHPUnit\Framework\TestCase
 {
-  private $model; 
+  private $model;
   private $concept;
   private $vocab;
   private $props;
 
   protected function setUp() {
-    require_once 'testconfig.inc';
+    putenv("LANGUAGE=en_GB.utf8");
     putenv("LC_ALL=en_GB.utf8");
     setlocale(LC_ALL, 'en_GB.utf8');
     bindtextdomain('skosmos', 'resource/translations');
     bind_textdomain_codeset('skosmos', 'UTF-8');
     textdomain('skosmos');
 
-    $this->model = new Model(new GlobalConfig('/../tests/testconfig.inc'));
+    $this->model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
     $this->vocab = $this->model->getVocabulary('mapping');
     $concepts = $this->vocab->getConceptInfo('http://www.skosmos.skos/mapping/m1', 'en');
     $this->concept = $concepts[0];
     $this->props = $this->concept->getMappingProperties();
   }
-  
+
   /**
    * @covers ConceptMappingPropertyValue::__construct
    */
@@ -125,7 +125,7 @@ class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
     $this->assertInstanceOf('Vocabulary', $propvals['Eelhttp://www.skosmos.skos/test/ta115']->getExVocab());
     $this->assertEquals('test', $propvals['Eelhttp://www.skosmos.skos/test/ta115']->getExVocab()->getId());
   }
-  
+
   /**
    * @covers ConceptMappingPropertyValue::getVocabName
    */
@@ -133,7 +133,7 @@ class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
     $propvals = $this->props['skos:exactMatch']->getValues();
     $this->assertEquals('Test ontology', $propvals['Eelhttp://www.skosmos.skos/test/ta115']->getVocabName());
   }
-  
+
   /**
    * @covers ConceptMappingPropertyValue::getUri
    */
@@ -141,7 +141,7 @@ class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
     $propvals = $this->props['skos:exactMatch']->getValues();
     $this->assertEquals('http://www.skosmos.skos/test/ta115', $propvals['Eelhttp://www.skosmos.skos/test/ta115']->getUri());
   }
-  
+
   /**
    * @covers ConceptMappingPropertyValue::getVocab
    */
@@ -149,7 +149,7 @@ class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
     $propvals = $this->props['skos:exactMatch']->getValues();
     $this->assertEquals($this->vocab, $propvals['Eelhttp://www.skosmos.skos/test/ta115']->getVocab());
   }
-  
+
   /**
    * @covers ConceptMappingPropertyValue::getType
    */
@@ -157,7 +157,7 @@ class ConceptMappingPropertyValueTest extends PHPUnit_Framework_TestCase
     $propvals = $this->props['skos:exactMatch']->getValues();
     $this->assertEquals('skos:exactMatch', $propvals['Eelhttp://www.skosmos.skos/test/ta115']->getType());
   }
-  
+
   /**
    * @covers ConceptMappingPropertyValue::__toString
    */
