@@ -1,20 +1,20 @@
 <?php
 
-class ConceptPropertyValueLiteralTest extends PHPUnit_Framework_TestCase
+class ConceptPropertyValueLiteralTest extends PHPUnit\Framework\TestCase
 {
-  private $model; 
+  private $model;
   private $concept;
   private $vocab;
 
   protected function setUp() {
-    require_once 'testconfig.inc';
+    putenv("LANGUAGE=en_GB.utf8");
     putenv("LC_ALL=en_GB.utf8");
     setlocale(LC_ALL, 'en_GB.utf8');
     bindtextdomain('skosmos', 'resource/translations');
     bind_textdomain_codeset('skosmos', 'UTF-8');
     textdomain('skosmos');
 
-    $this->model = new Model(new GlobalConfig('/../tests/testconfig.inc'));
+    $this->model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
     $this->vocab = $this->model->getVocabulary('test');
     $results = $this->vocab->getConceptInfo('http://www.skosmos.skos/test/ta112', 'en');
     $this->concept = reset($results);
@@ -25,7 +25,8 @@ class ConceptPropertyValueLiteralTest extends PHPUnit_Framework_TestCase
    */
   public function testConstructor() {
     $litmock = $this->getMockBuilder('EasyRdf\Literal')->disableOriginalConstructor()->getMock();
-    $prop = new ConceptPropertyValueLiteral($litmock, 'skosmos:someProperty');
+    $resmock = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
+    $prop = new ConceptPropertyValueLiteral($this->model, $this->vocab, $resmock, $litmock, 'skosmos:someProperty');
     $this->assertEquals(null, $prop->__toString());
   }
 
@@ -51,7 +52,7 @@ class ConceptPropertyValueLiteralTest extends PHPUnit_Framework_TestCase
 
   /**
    * @covers ConceptPropertyValueLiteral::getLabel
-   * @expectedException PHPUnit_Framework_Error
+   * @expectedException PHPUnit\Framework\Error\Warning
    */
   public function testGetLabelThatIsABrokenDate() {
     $vocab = $this->model->getVocabulary('dates');
@@ -101,7 +102,8 @@ class ConceptPropertyValueLiteralTest extends PHPUnit_Framework_TestCase
    */
   public function testToStringEmpty() {
     $litmock = $this->getMockBuilder('EasyRdf\Literal')->disableOriginalConstructor()->getMock();
-    $lit = new ConceptPropertyValueLiteral($litmock, 'skosmos:testType');
+    $resmock = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
+    $lit = new ConceptPropertyValueLiteral($this->model, $this->vocab, $resmock, $litmock, 'skosmos:testType');
     $this->assertEquals('', $lit);
   }
 
@@ -110,7 +112,8 @@ class ConceptPropertyValueLiteralTest extends PHPUnit_Framework_TestCase
    */
   public function testGetNotation() {
     $litmock = $this->getMockBuilder('EasyRdf\Literal')->disableOriginalConstructor()->getMock();
-    $lit = new ConceptPropertyValueLiteral($litmock, 'skosmos:testType');
+    $resmock = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
+    $lit = new ConceptPropertyValueLiteral($this->model, $this->vocab, $resmock, $litmock, 'skosmos:testType');
     $this->assertEquals(null, $lit->getNotation());
   }
 
@@ -120,27 +123,30 @@ class ConceptPropertyValueLiteralTest extends PHPUnit_Framework_TestCase
   public function testGetContainsHtmlWhenThereIsNone() {
     $litmock = $this->getMockBuilder('EasyRdf\Literal')->disableOriginalConstructor()->getMock();
     $litmock->method('getValue')->will($this->returnValue('a regular literal'));
-    $lit = new ConceptPropertyValueLiteral($litmock, 'skosmos:testType');
+    $resmock = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
+    $lit = new ConceptPropertyValueLiteral($this->model, $this->vocab, $resmock, $litmock, 'skosmos:testType');
     $this->assertFalse($lit->getContainsHtml());
   }
-  
+
   /**
    * @covers ConceptPropertyValueLiteral::getContainsHtml
    */
   public function testGetContainsHtmlWhenThereIsOnlyAOpeningTag() {
     $litmock = $this->getMockBuilder('EasyRdf\Literal')->disableOriginalConstructor()->getMock();
     $litmock->method('getValue')->will($this->returnValue('a <a href=\"http://skosmos.org\"> literal with broken html'));
-    $lit = new ConceptPropertyValueLiteral($litmock, 'skosmos:testType');
+    $resmock = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
+    $lit = new ConceptPropertyValueLiteral($this->model, $this->vocab, $resmock, $litmock, 'skosmos:testType');
     $this->assertFalse($lit->getContainsHtml());
   }
-  
+
   /**
    * @covers ConceptPropertyValueLiteral::getContainsHtml
    */
   public function testGetContainsHtml() {
     $litmock = $this->getMockBuilder('EasyRdf\Literal')->disableOriginalConstructor()->getMock();
     $litmock->method('getValue')->will($this->returnValue('a <a href=\"http://skosmos.org\">literal</a> with valid html'));
-    $lit = new ConceptPropertyValueLiteral($litmock, 'skosmos:testType');
+    $resmock = $this->getMockBuilder('EasyRdf\Resource')->disableOriginalConstructor()->getMock();
+    $lit = new ConceptPropertyValueLiteral($this->model, $this->vocab, $resmock, $litmock, 'skosmos:testType');
     $this->assertTrue($lit->getContainsHtml());
   }
 }
