@@ -1,19 +1,20 @@
 <?php
 
-class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
+class ConceptPropertyValueTest extends PHPUnit\Framework\TestCase
 {
-  private $model; 
+  private $model;
   private $concept;
   private $vocab;
-    
+
   protected function setUp() {
+    putenv("LANGUAGE=en_GB.utf8");
     putenv("LC_ALL=en_GB.utf8");
     setlocale(LC_ALL, 'en_GB.utf8');
     bindtextdomain('skosmos', 'resource/translations');
     bind_textdomain_codeset('skosmos', 'UTF-8');
     textdomain('skosmos');
 
-    $this->model = new Model(new GlobalConfig('/../tests/testconfig.inc'));
+    $this->model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
     $this->vocab = $this->model->getVocabulary('test');
     $results = $this->vocab->getConceptInfo('http://www.skosmos.skos/test/ta112', 'en');
     $this->concept = reset($results);
@@ -57,7 +58,7 @@ class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
     $this->assertEquals('english lit', $mapping->getLabel('en'));
     $this->assertEquals('default lit', $mapping->getLabel());
   }
-  
+
   /**
    * @covers ConceptPropertyValue::getType
    */
@@ -66,7 +67,7 @@ class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
     $propvals = $props['skos:narrower']->getValues();
     $this->assertEquals('skos:narrower', $propvals['Crucian carphttp://www.skosmos.skos/test/ta121']->getType());
   }
-  
+
   /**
    * @covers ConceptPropertyValue::getUri
    */
@@ -75,7 +76,7 @@ class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
     $propvals = $props['skos:narrower']->getValues();
     $this->assertEquals('http://www.skosmos.skos/test/ta121', $propvals['Crucian carphttp://www.skosmos.skos/test/ta121']->getUri());
   }
-  
+
   /**
    * @covers ConceptPropertyValue::getVocab
    */
@@ -114,7 +115,7 @@ class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
     $propval = new ConceptPropertyValue($this->model, $this->vocab, $mockres, 'en');
     $this->assertEquals(null, $propval->getNotation());
   }
-  
+
   /**
    * @covers ConceptPropertyValue::__toString
    */
@@ -125,7 +126,7 @@ class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
     $propvals = $props['skos:broader']->getValues();
     $this->assertEquals('Carp', (string)$propvals['665Carphttp://www.skosmos.skos/test/ta112']);
   }
-  
+
   /**
    * @covers ConceptPropertyValue::__toString
    */
@@ -144,7 +145,7 @@ class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
     $propval = new ConceptPropertyValue($this->model, $mockvoc, $mockres, null);
     $this->assertEquals('T3STTerm label', (string)$propval);
   }
-   
+
   /**
    * @covers ConceptPropertyValue::__toString
    */
@@ -200,5 +201,15 @@ class ConceptPropertyValueTest extends PHPUnit_Framework_TestCase
     $mockres = $this->getMockBuilder('EasyRdf\\Resource')->disableOriginalConstructor()->getMock();
     $propval = new ConceptPropertyValue($this->model, $this->vocab, $mockres, 'en');
     $this->assertEquals(null, $propval->getSubMembers());
+  }
+
+  public function testGetReifiedPropertyValues() {
+    $vocab = $this->model->getVocabulary('xl');
+    $concept = $vocab->getConceptInfo('http://www.skosmos.skos/xl/c1', 'en')[0];
+    $props = $concept->getProperties();
+    $vals = $props['skos:definition']->getValues();
+    $val = reset($vals);
+    $reified_vals = $val->getReifiedPropertyValues();
+    $this->assertEquals(2, count($reified_vals));
   }
 }
