@@ -1,36 +1,37 @@
 <?php
 
-class GenericSparqlTest extends PHPUnit_Framework_TestCase
+class GenericSparqlTest extends PHPUnit\Framework\TestCase
 {
-  private $model; 
-  private $graph; 
+  private $model;
+  private $graph;
   private $sparql;
   private $vocab;
   private $params;
 
   protected function setUp() {
+    putenv("LANGUAGE=en_GB.utf8");
     putenv("LC_ALL=en_GB.utf8");
     setlocale(LC_ALL, 'en_GB.utf8');
-    $this->model = new Model(new GlobalConfig('/../tests/testconfig.inc'));
+    $this->model = new Model(new GlobalConfig('/../tests/testconfig.ttl'));
     $this->vocab = $this->model->getVocabulary('test');
     $this->graph = $this->vocab->getGraph();
     $this->params = $this->getMockBuilder('ConceptSearchParameters')->disableOriginalConstructor()->getMock();
-    $this->sparql = new GenericSparql('http://localhost:3030/ds/sparql', $this->graph, $this->model);
+    $this->sparql = new GenericSparql('http://localhost:13030/ds/sparql', $this->graph, $this->model);
   }
- 
+
   /**
    * @covers GenericSparql::__construct
    */
   public function testConstructor() {
-    $gs = new GenericSparql('http://localhost:3030/ds/sparql', $this->graph, $this->model);
+    $gs = new GenericSparql('http://localhost:13030/ds/sparql', $this->graph, $this->model);
     $this->assertInstanceOf('GenericSparql', $gs);
   }
-  
+
   /**
    * @covers GenericSparql::getGraph
    */
   public function testGetGraph() {
-    $gs = new GenericSparql('http://localhost:3030/ds/sparql', $this->graph, $this->model);
+    $gs = new GenericSparql('http://localhost:13030/ds/sparql', $this->graph, $this->model);
     $this->assertEquals($this->graph, $gs->getGraph());
   }
 
@@ -43,7 +44,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $actual = $this->sparql->countConcepts();
     $this->assertEquals(16, $actual['http://www.w3.org/2004/02/skos/core#Concept']['count']);
   }
-  
+
   /**
    * @covers GenericSparql::countLangConcepts
    * @covers GenericSparql::generateCountLangConceptsQuery
@@ -54,7 +55,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $this->assertEquals(11, $actual['en']['skos:prefLabel']);
     $this->assertEquals(1, $actual['en']['skos:altLabel']);
   }
-  
+
   /**
    * @covers GenericSparql::countLangConcepts
    * @covers GenericSparql::generateCountLangConceptsQuery
@@ -66,7 +67,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $this->assertEquals(1, $actual['en']['skos:altLabel']);
     $this->assertEquals(2, $actual['fi']['skos:prefLabel']);
   }
-  
+
   /**
    * @covers GenericSparql::queryFirstCharacters
    * @covers GenericSparql::generateFirstCharactersQuery
@@ -85,7 +86,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $actual = $this->sparql->queryLabel('http://notfound', null);
     $this->assertEquals(null, $actual);
   }
-  
+
   /**
    * @covers GenericSparql::queryLabel
    */
@@ -93,7 +94,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $actual = $this->sparql->queryLabel('http://www.skosmos.skos/test/ta120', null);
     $this->assertEquals(array(), $actual);
   }
-  
+
   /**
    * @covers GenericSparql::queryConceptsAlphabetical
    * @covers GenericSparql::generateAlphabeticalListQuery
@@ -124,7 +125,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     );
     $this->assertEquals($expected, $actual);
   }
-  
+
   /**
    * @covers GenericSparql::queryConceptsAlphabetical
    * @covers GenericSparql::generateAlphabeticalListQuery
@@ -150,7 +151,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     );
     $this->assertEquals($expected, $actual);
   }
-  
+
   /**
    * @covers GenericSparql::queryConceptsAlphabetical
    * @covers GenericSparql::generateAlphabeticalListQuery
@@ -176,7 +177,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     );
     $this->assertEquals($expected, $actual);
   }
-  
+
   /**
    * @covers GenericSparql::queryConceptsAlphabetical
    * @covers GenericSparql::generateAlphabeticalListQuery
@@ -200,7 +201,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $this->assertEquals(1, sizeof($actual));
     $this->assertEquals('-"special" character \\example\\', $actual[0]['prefLabel']);
   }
-  
+
   /**
    * @covers GenericSparql::queryConceptsAlphabetical
    * @covers GenericSparql::generateAlphabeticalListQuery
@@ -212,7 +213,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $this->assertEquals(1, sizeof($actual));
     $this->assertContains("3D", $actual[0]['prefLabel']);
   }
-  
+
   /**
    * @covers GenericSparql::queryConceptsAlphabetical
    * @covers GenericSparql::generateAlphabeticalListQuery
@@ -235,7 +236,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
    */
   public function testQueryConceptInfoWithMultipleVocabs()
   {
-    $this->sparql = new GenericSparql('http://localhost:3030/ds/sparql', '?graph', $this->model);
+    $this->sparql = new GenericSparql('http://localhost:13030/ds/sparql', '?graph', $this->model);
     $voc2 = $this->model->getVocabulary('test');
     $voc3 = $this->model->getVocabulary('dates');
     $voc4 = $this->model->getVocabulary('groups');
@@ -257,7 +258,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
    */
   public function testQueryConceptInfoWithAllVocabs()
   {
-    $this->sparql = new GenericSparql('http://localhost:3030/ds/sparql', '?graph', $this->model);
+    $this->sparql = new GenericSparql('http://localhost:13030/ds/sparql', '?graph', $this->model);
     $actual = $this->sparql->queryConceptInfo(array('http://www.skosmos.skos/test/ta121', 'http://www.skosmos.skos/groups/ta111'), null, null, 'en');
     $this->assertInstanceOf('Concept', $actual[0]);
     $this->assertEquals('http://www.skosmos.skos/test/ta121', $actual[0]->getUri());
@@ -334,7 +335,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
   {
     $actual = $this->sparql->queryConceptScheme('http://www.skosmos.skos/test/conceptscheme');
     $this->assertInstanceOf('EasyRdf\Graph', $actual);
-    $this->assertEquals('http://localhost:3030/ds/sparql', $actual->getUri());
+    $this->assertEquals('http://localhost:13030/ds/sparql', $actual->getUri());
   }
 
   /**
@@ -349,6 +350,49 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
       $this->assertEquals('http://www.skosmos.skos/test/conceptscheme', $scheme);
       $this->assertEquals('Test conceptscheme', $label['label']);
     }
+  }
+
+  /**
+   * @covers GenericSparql::queryConceptSchemes
+   * @covers GenericSparql::generateQueryConceptSchemesQuery
+   * @covers GenericSparql::transformQueryConceptSchemesResults
+   */
+  public function testQueryConceptSchemesSubject()
+  {
+      $sparql = new GenericSparql('http://localhost:13030/ds/sparql', 'http://www.skosmos.skos/test-concept-schemes/', $this->model);
+
+      $actual = $sparql->queryConceptSchemes('en');
+      $expected = array(
+          'http://exemple.fr/domains' => array(
+              'prefLabel' => 'Special Domains Concept Scheme'
+          ),
+          'http://exemple.fr/mt1' => array(
+              'prefLabel' => 'Micro-Thesaurus 1',
+              'subject' => array(
+                  'uri' => 'http://exemple.fr/d1',
+                  'prefLabel' => 'Domain 1'
+              )
+          ),
+          'http://exemple.fr/mt2' => array(
+              'prefLabel' => 'Micro-Thesaurus 2',
+              'subject' => array(
+                  'uri' => 'http://exemple.fr/d1',
+                  'prefLabel' => 'Domain 1'
+              )
+          ),
+          'http://exemple.fr/mt3' => array(
+              'prefLabel' => 'Micro-Thesaurus 3',
+              'subject' => array(
+                  'uri' => 'http://exemple.fr/d2',
+                  'prefLabel' => 'Domain 2'
+              )
+          ),
+          'http://exemple.fr/thesaurus' => array(
+              'prefLabel' => 'The Thesaurus'
+          ),
+      );
+
+      $this->assertEquals($expected, $actual);
   }
 
   /**
@@ -367,7 +411,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $voc2 = $this->model->getVocabulary('groups');
     $this->params->method('getSearchTerm')->will($this->returnValue('Carp'));
     $this->params->method('getVocabs')->will($this->returnValue(array($voc, $voc2)));
-    $sparql = new GenericSparql('http://localhost:3030/ds/sparql', '?graph', $this->model);
+    $sparql = new GenericSparql('http://localhost:13030/ds/sparql', '?graph', $this->model);
     $actual = $sparql->queryConcepts(array($voc, $voc2), null, null, $this->params);
     $this->assertEquals(2, sizeof($actual));
     $this->assertEquals('http://www.skosmos.skos/groups/ta112', $actual[0]['uri']);
@@ -390,14 +434,14 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
       // returns 3 concepts without the scheme limit, and only 2 with the scheme limit below
       $this->params->method('getSearchTerm')->will($this->returnValue('concept*'));
       $this->params->method('getSchemeLimit')->will($this->returnValue(array('http://www.skosmos.skos/multiple-schemes/cs1', 'http://www.skosmos.skos/multiple-schemes/cs2')));
-      $sparql = new GenericSparql('http://localhost:3030/ds/sparql', 'http://www.skosmos.skos/multiple-schemes/', $this->model);
+      $sparql = new GenericSparql('http://localhost:13030/ds/sparql', 'http://www.skosmos.skos/multiple-schemes/', $this->model);
       $actual = $sparql->queryConcepts(array($voc), null, null, $this->params);
       $this->assertEquals(2, sizeof($actual));
       $this->assertEquals('http://www.skosmos.skos/multiple-schemes/c1-in-cs1', $actual[0]['uri']);
       $this->assertEquals('http://www.skosmos.skos/multiple-schemes/c2-in-cs2', $actual[1]['uri']);
   }
-  
-  
+
+
   /**
    * @covers GenericSparql::queryConcepts
    * @covers GenericSparql::generateConceptSearchQueryCondition
@@ -515,7 +559,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $this->assertEquals(1, sizeof($actual));
     $this->assertEquals('Bass', $actual[0]['prefLabel']);
   }
-  
+
   /**
    * @covers GenericSparql::queryConcepts
    * @covers GenericSparql::generateConceptSearchQueryCondition
@@ -534,7 +578,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     foreach($actual as $match)
       $this->assertContains('bass', $match['prefLabel'], '',true);
   }
-  
+
   /**
    * @covers GenericSparql::queryConcepts
    * @covers GenericSparql::generateConceptSearchQueryCondition
@@ -553,7 +597,24 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     foreach($actual as $match)
       $this->assertContains('bass', $match['prefLabel'], '',true);
   }
-  
+
+  /**
+   * @covers GenericSparql::queryConcepts
+   * @covers GenericSparql::generateConceptSearchQueryCondition
+   * @covers GenericSparql::generateConceptSearchQueryInner
+   * @covers GenericSparql::generateConceptSearchQuery
+   * @covers GenericSparql::transformConceptSearchResults
+   * @covers GenericSparql::transformConceptSearchResult
+   * @covers GenericSparql::shortenUri
+   */
+  public function testQueryConceptsDoubleQuotesTerm()
+  {
+    $voc = $this->model->getVocabulary('test');
+    $this->params->method('getSearchTerm')->will($this->returnValue('"'));
+    $actual = $this->sparql->queryConcepts(array($voc), null, null, $this->params);
+    $this->assertEquals(0, sizeof($actual));
+  }
+
   /**
    * @covers GenericSparql::queryLabel
    * @covers GenericSparql::generateLabelQuery
@@ -574,7 +635,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $expected = array('en' => 'Carp');
     $this->assertEquals($expected, $actual);
   }
-  
+
   /**
    * @covers GenericSparql::queryLabel
    * @covers GenericSparql::generateLabelQuery
@@ -597,7 +658,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $expected = array('http://www.skosmos.skos/test/ta1' => array('label' => 'Fish'));
     $this->assertEquals($expected, $actual);
   }
-  
+
   /**
    * @covers GenericSparql::queryProperty
    * @covers GenericSparql::transformPropertyQueryResults
@@ -609,7 +670,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     $expected = array();
     $this->assertEquals($expected, $actual);
   }
-  
+
   /**
    * @covers GenericSparql::queryProperty
    * @covers GenericSparql::transformPropertyQueryResults
@@ -630,22 +691,22 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
   {
     $actual = $this->sparql->queryTransitiveProperty('http://www.skosmos.skos/test/ta111', array('skos:broader'), 'en', '10');
     $expected = array(
-      'http://www.skosmos.skos/test/ta111' => 
+      'http://www.skosmos.skos/test/ta111' =>
         array(
           'label' => 'Tuna',
-          'direct' => 
+          'direct' =>
           array (
             0 => 'http://www.skosmos.skos/test/ta1',
           ),
         ),
-        'http://www.skosmos.skos/test/ta1' => 
+        'http://www.skosmos.skos/test/ta1' =>
         array (
           'label' => 'Fish',
         )
     );
     $this->assertEquals($expected, $actual);
   }
-  
+
   /**
    * @covers GenericSparql::queryTransitiveProperty
    * @covers GenericSparql::generateTransitivePropertyQuery
@@ -655,22 +716,22 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
   {
     $actual = $this->sparql->queryTransitiveProperty('http://www.skosmos.skos/test/ta122', array('skos:broader'), 'en', '10');
     $expected = array(
-      'http://www.skosmos.skos/test/ta122' => 
+      'http://www.skosmos.skos/test/ta122' =>
         array (
           'label' => 'Black sea bass',
-          'direct' => 
+          'direct' =>
           array (
             0 => 'http://www.skosmos.skos/test/ta116',
           ),
         ),
-        'http://www.skosmos.skos/test/ta1' => 
+        'http://www.skosmos.skos/test/ta1' =>
         array (
           'label' => 'Fish',
         ),
-        'http://www.skosmos.skos/test/ta116' => 
+        'http://www.skosmos.skos/test/ta116' =>
         array (
           'label' => 'Bass',
-          'direct' => 
+          'direct' =>
           array (
             0 => 'http://www.skosmos.skos/test/ta1',
           ),
@@ -695,7 +756,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     foreach ($expected as $uri)
       $this->assertArrayHasKey($uri, $actual_uris);
   }
-  
+
   /**
    * @covers GenericSparql::queryChildren
    * @covers GenericSparql::generateChildQuery
@@ -746,7 +807,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
       'http://www.skosmos.skos/test/ta116' => array (
         'uri' => 'http://www.skosmos.skos/test/ta116',
         'prefLabel' => 'Bass',
-        'broader' => 
+        'broader' =>
         array (
           0 => 'http://www.skosmos.skos/test/ta1',
         ),
@@ -754,7 +815,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
       'http://www.skosmos.skos/test/ta122' => array (
         'uri' => 'http://www.skosmos.skos/test/ta122',
         'prefLabel' => 'Black sea bass',
-        'broader' => 
+        'broader' =>
         array (
           0 => 'http://www.skosmos.skos/test/ta116',
         ),
@@ -762,7 +823,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
     );
     $props = array (
       'uri' => 'http://www.skosmos.skos/test/ta1',
-      'top' => 'http://www.skosmos.skos/test/conceptscheme',
+      'tops' => array('http://www.skosmos.skos/test/conceptscheme'),
       'prefLabel' => 'Fish',
     );
     $narrowers = array (
@@ -833,7 +894,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
   {
     $voc = $this->model->getVocabulary('groups');
     $graph = $voc->getGraph();
-    $sparql = new GenericSparql('http://localhost:3030/ds/sparql', $graph, $this->model);
+    $sparql = new GenericSparql('http://localhost:13030/ds/sparql', $graph, $this->model);
     $actual = $sparql->ListConceptGroups('http://www.w3.org/2004/02/skos/core#Collection', 'en', false);
     $expected = array (0 => array ('prefLabel' => 'Fish', 'uri' => 'http://www.skosmos.skos/groups/fish', 'hasMembers' => true, 'childGroups' => array('http://www.skosmos.skos/groups/sub')), 1 => array ('prefLabel' => 'Freshwater fish', 'uri' => 'http://www.skosmos.skos/groups/fresh', 'hasMembers' => true), 2 => array ('prefLabel' => 'Saltwater fish', 'uri' => 'http://www.skosmos.skos/groups/salt', 'hasMembers' => true),3 => array ('prefLabel' => 'Submarine-like fish', 'uri' => 'http://www.skosmos.skos/groups/sub', 'hasMembers' => true));
     $this->assertEquals($expected, $actual);
@@ -848,12 +909,12 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
   {
     $voc = $this->model->getVocabulary('groups');
     $graph = $voc->getGraph();
-    $sparql = new GenericSparql('http://localhost:3030/ds/sparql', $graph, $this->model);
+    $sparql = new GenericSparql('http://localhost:13030/ds/sparql', $graph, $this->model);
     $actual = $sparql->ListConceptGroupContents('http://www.w3.org/2004/02/skos/core#Collection', 'http://www.skosmos.skos/groups/salt', 'en');
     $this->assertEquals('http://www.skosmos.skos/groups/ta113', $actual[0]['uri']);
     $this->assertEquals(1, sizeof($actual));
   }
-  
+
   /**
    * @covers GenericSparql::listConceptGroupContents
    * @covers GenericSparql::generateConceptGroupContentsQuery
@@ -863,7 +924,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
   {
       $voc = $this->model->getVocabulary('showDeprecated');
       $graph = $voc->getGraph();
-      $sparql = new GenericSparql('http://localhost:3030/ds/sparql', $graph, $this->model);
+      $sparql = new GenericSparql('http://localhost:13030/ds/sparql', $graph, $this->model);
       $actual = $sparql->ListConceptGroupContents('http://www.w3.org/2004/02/skos/core#Collection', 'http://www.skosmos.skos/groups/salt', 'en', $voc->getConfig()->getShowDeprecated());
       $expected = array (
           0 => array (
@@ -894,7 +955,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
   {
     $voc = $this->model->getVocabulary('changes');
     $graph = $voc->getGraph();
-    $sparql = new GenericSparql('http://localhost:3030/ds/sparql', $graph, $this->model);
+    $sparql = new GenericSparql('http://localhost:13030/ds/sparql', $graph, $this->model);
     $actual = $sparql->queryChangeList('en', 0, 'dc11:created');
     $order = array();
     foreach($actual as $concept) {
@@ -912,7 +973,7 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
   {
     $voc = $this->model->getVocabulary('test');
     $graph = $voc->getGraph();
-    $sparql = new GenericSparql('http://localhost:3030/ds/sparql', $graph, $this->model);
+    $sparql = new GenericSparql('http://localhost:13030/ds/sparql', $graph, $this->model);
     $this->params->method('getSearchTerm')->will($this->returnValue('*'));
     $this->params->method('getTypeLimit')->will($this->returnValue(array('mads:Topic')));
     $actual = $this->sparql->queryConcepts(array($voc), null, true, $this->params);
@@ -945,5 +1006,17 @@ class GenericSparqlTest extends PHPUnit_Framework_TestCase
 );
     $this->assertEquals(array('en' => 'Bass'), $actual[0]['prefLabels']);
     $this->assertEquals(array(0 => array('uri' => 'http://www.skosmos.skos/test/ta1')), $actual[0]['skos:broader']);
+  }
+
+  /**
+   * @covers GenericSparql::querySuperProperties
+   */
+  public function testQuerySuperProperties()
+  {
+      $this->sparql = new GenericSparql('http://localhost:13030/ds/sparql', '?graph', $this->model);
+      $actual = $this->sparql->querySuperProperties('http://example.com/myns#property');
+      $this->assertEquals(1, sizeof($actual));
+      $expected = array('http://example.com/myns#superProperty');
+      $this->assertEquals($actual, $expected);
   }
 }
