@@ -1,36 +1,37 @@
 <?php
 
-class RequestTest extends PHPUnit_Framework_TestCase
+class RequestTest extends PHPUnit\Framework\TestCase
 {
   private $model;
   private $request;
-  
+
   protected function setUp() {
+    putenv("LANGUAGE=en_GB.utf8");
     putenv("LC_ALL=en_GB.utf8");
     setlocale(LC_ALL, 'en_GB.utf8');
     bindtextdomain('skosmos', 'resource/translations');
     bind_textdomain_codeset('skosmos', 'UTF-8');
     textdomain('skosmos');
 
-    $config = new GlobalConfig('/../tests/testconfig.inc');
+    $config = new GlobalConfig('/../tests/testconfig.ttl');
     $this->model = new Model($config);
     $this->request = new Request($this->model);
   }
-  
+
   /**
    * @covers Request::__construct
    */
   public function testConstructor() {
       $this->assertInstanceOf('Request', new Request($this->model));
   }
-  
+
   /**
    * @covers Request::getQueryParamBoolean
    */
   public function testGetQueryParamBooleanDefaultValue() {
       $this->assertTrue($this->request->getQueryParamBoolean('notfoundatall', true));
   }
-  
+
   /**
    * @covers Request::getVocabList
    */
@@ -42,7 +43,7 @@ class RequestTest extends PHPUnit_Framework_TestCase
       }
     }
   }
-  
+
   /**
    * @covers Request::setVocabids
    * @covers Request::getVocabid
@@ -51,7 +52,7 @@ class RequestTest extends PHPUnit_Framework_TestCase
     $this->request->setVocabids(array('testing', 'setter'));
     $this->assertEquals(array('testing', 'setter'), $this->request->getVocabid());
   }
-  
+
   /**
    * @covers Request::setVocab
    * @covers Request::getVocab
@@ -60,7 +61,7 @@ class RequestTest extends PHPUnit_Framework_TestCase
     $this->request->setVocab('test');
     $this->assertInstanceOf('Vocabulary', $this->request->getVocab());
   }
-  
+
   /**
    * @covers Request::setVocab
    * @covers Request::getVocabid
@@ -69,7 +70,7 @@ class RequestTest extends PHPUnit_Framework_TestCase
     $this->request->setVocab('test dates');
     $this->assertEquals('test dates', $this->request->getVocabid());
   }
-  
+
   /**
    * @covers Request::getVocabid
    */
@@ -77,7 +78,7 @@ class RequestTest extends PHPUnit_Framework_TestCase
     $this->request->setVocab('test');
     $this->assertEquals('test', $this->request->getVocabId());
   }
-  
+
   /**
    * @covers Request::setUri
    * @covers Request::getUri
@@ -87,7 +88,7 @@ class RequestTest extends PHPUnit_Framework_TestCase
     $this->request->setUri('www.skosmos.org');
     $this->assertEquals('www.skosmos.org', $this->request->getUri());
   }
-  
+
   /**
    * @covers Request::setContentLang
    * @covers Request::getContentLang
@@ -98,7 +99,7 @@ class RequestTest extends PHPUnit_Framework_TestCase
     $clang = $this->request->setContentLang('en');
     $this->assertEquals('en', $this->request->getContentLang());
   }
-  
+
   /**
    * @covers Request::setContentLang
    * @covers Request::verifyContentLang
@@ -107,7 +108,7 @@ class RequestTest extends PHPUnit_Framework_TestCase
     $clang = $this->request->setContentLang('fi');
     $this->assertEquals('fi', $this->request->getContentLang());
   }
-  
+
   /**
    * @covers Request::setContentLang
    * @covers Request::getContentLang
@@ -118,7 +119,7 @@ class RequestTest extends PHPUnit_Framework_TestCase
     $clang = $this->request->setContentLang('ru');
     $this->assertEquals('en', $this->request->getContentLang());
   }
-  
+
   /**
    * @covers Request::setLang
    * @covers Request::getLang
@@ -128,15 +129,15 @@ class RequestTest extends PHPUnit_Framework_TestCase
     $clang = $this->request->setLang('en');
     $this->assertEquals('en', $this->request->getLang());
   }
-  
+
   /**
    * @covers Request::getLetter
    */
   public function testGetLetterWhenNotSet() {
     $this->request->setVocab('test');
-    $this->assertEquals('A', $this->request->getLetter());
+    $this->assertEquals('', $this->request->getLetter());
   }
-  
+
   /**
    * @covers Request::setLetter
    * @covers Request::getLetter
@@ -146,7 +147,7 @@ class RequestTest extends PHPUnit_Framework_TestCase
     $this->request->setLetter('X');
     $this->assertEquals('X', $this->request->getLetter());
   }
-  
+
   /**
    * @covers Request::setPage
    * @covers Request::getPage
@@ -155,5 +156,5 @@ class RequestTest extends PHPUnit_Framework_TestCase
     $this->request->setPage('index');
     $this->assertEquals('index', $this->request->getPage());
   }
-  
+
 }
