@@ -29,7 +29,7 @@ class WebController extends Controller
         // initialize Twig templates
         $tmpDir = $model->getConfig()->getTemplateCache();
 
-        // check if the cache pointed by config.inc exists, if not we create it.
+        // check if the cache pointed by config.ttl exists, if not we create it.
         if (!file_exists($tmpDir)) {
             mkdir($tmpDir);
         }
@@ -41,14 +41,10 @@ class WebController extends Controller
         $this->twig->addExtension(new Twig_Extensions_Extension_I18n());
         // used for setting the base href for the relative urls
         $this->twig->addGlobal("BaseHref", $this->getBaseHref());
-        // setting the service name string from the config.inc
+        // setting the service name string from the config.ttl
         $this->twig->addGlobal("ServiceName", $this->model->getConfig()->getServiceName());
-        // setting the service logo location from the config.inc
-        if ($this->model->getConfig()->getServiceLogo() !== null) {
-            $this->twig->addGlobal("ServiceLogo", $this->model->getConfig()->getServiceLogo());
-        }
 
-        // setting the service custom css file from the config.inc
+        // setting the service custom css file from the config.ttl
         if ($this->model->getConfig()->getCustomCss() !== null) {
             $this->twig->addGlobal("ServiceCustomCss", $this->model->getConfig()->getCustomCss());
         }
@@ -152,7 +148,8 @@ class WebController extends Controller
         header('Vary: Accept-Language'); // inform caches that a decision was made based on Accept header
         $this->negotiator = new \Negotiation\LanguageNegotiator();
         $langcodes = array_keys($this->languages);
-        $acceptLanguage = filter_input(INPUT_SERVER, 'HTTP_ACCEPT_LANGUAGE', FILTER_SANITIZE_STRING) ? filter_input(INPUT_SERVER, 'HTTP_ACCEPT_LANGUAGE', FILTER_SANITIZE_STRING) : '';
+        // using a random language from the configured UI languages when there is no accept language header set
+        $acceptLanguage = filter_input(INPUT_SERVER, 'HTTP_ACCEPT_LANGUAGE', FILTER_SANITIZE_STRING) ? filter_input(INPUT_SERVER, 'HTTP_ACCEPT_LANGUAGE', FILTER_SANITIZE_STRING) : $langcodes[0];
         $bestLang = $this->negotiator->getBest($acceptLanguage, $langcodes);
         if (isset($bestLang) && in_array($bestLang, $langcodes)) {
             return $bestLang->getValue();
@@ -163,7 +160,7 @@ class WebController extends Controller
     }
 
     /**
-     * Determines a css class that controls width and positioning of the vocabulary list element. 
+     * Determines a css class that controls width and positioning of the vocabulary list element.
      * The layout is wider if the left/right box templates have not been provided.
      * @return string css class for the container eg. 'voclist-wide' or 'voclist-right'
      */
@@ -193,7 +190,7 @@ class WebController extends Controller
         $categoryLabel = $this->model->getClassificationLabel($request->getLang());
         $sortedVocabs = $this->model->getVocabularyList(false, true);
         $langList = $this->model->getLanguages($request->getLang());
-        $listStyle = $this->listStyle(); 
+        $listStyle = $this->listStyle();
 
         // render template
         echo $template->render(
@@ -225,7 +222,7 @@ class WebController extends Controller
             return;
         }
         $template = (in_array('skos:Concept', $results[0]->getType()) || in_array('skos:ConceptScheme', $results[0]->getType())) ? $this->twig->loadTemplate('concept-info.twig') : $this->twig->loadTemplate('group-contents.twig');
-        
+
         $crumbs = $vocab->getBreadCrumbs($request->getContentLang(), $uri);
         echo $template->render(array(
             'search_results' => $results,
@@ -257,7 +254,9 @@ class WebController extends Controller
         $feedbackName = $request->getQueryParamPOST('name');
         $feedbackEmail = $request->getQueryParamPOST('email');
         $feedbackVocab = $request->getQueryParamPOST('vocab');
-        $feedbackVocabEmail = ($vocab !== null) ? $vocab->getConfig()->getFeedbackRecipient() : null;
+
+        $feedbackVocabEmail = ($feedbackVocab !== null && $feedbackVocab !== '') ?
+            $this->model->getVocabulary($feedbackVocab)->getConfig()->getFeedbackRecipient() : null;
 
         // if the hidden field has been set a value we have found a spam bot
         // and we do not actually send the message.
@@ -276,15 +275,19 @@ class WebController extends Controller
             ));
     }
 
-    private function createFeedbackHeaders($fromName, $fromEmail, $toMail)
+    private function createFeedbackHeaders($fromName, $fromEmail, $toMail, $sender)
     {
         $headers = "MIME-Version: 1.0″ . '\r\n";
         $headers .= "Content-type: text/html; charset=UTF-8" . "\r\n";
-        if ($toMail) {
+        if (!empty($toMail)) {
             $headers .= "Cc: " . $this->model->getConfig()->getFeedbackAddress() . "\r\n";
         }
+        if (!empty($fromEmail)) {
+            $headers .= "Reply-To: $fromName <$fromEmail>\r\n";
+        }
 
-        $headers .= "From: $fromName <$fromEmail>" . "\r\n" . 'X-Mailer: PHP/' . phpversion();
+        $service = $this->model->getConfig()->getServiceName();
+        $headers .= "From: $fromName via $service <$sender>";
         return $headers;
     }
 
@@ -302,16 +305,23 @@ class WebController extends Controller
             $message = 'Feedback from vocab: ' . strtoupper($fromVocab) . "<br />" . $message;
         }
 
-        $subject = SERVICE_NAME . " feedback";
-        $headers = $this->createFeedbackHeaders($fromName, $fromEmail, $toMail);
-        $envelopeSender = FEEDBACK_ENVELOPE_SENDER;
+        $envelopeSender = $this->model->getConfig()->getFeedbackEnvelopeSender();
+        $subject = $this->model->getConfig()->getServiceName() . " feedback";
+        // determine the sender address of the message
+        $sender = $this->model->getConfig()->getFeedbackSender();
+        if (empty($sender)) $sender = $envelopeSender;
+        if (empty($sender)) $sender = $this->model->getConfig()->getFeedbackAddress();
+
+        // determine sender name - default to "anonymous user" if not given by user
+        if (empty($fromName)) $fromName = "anonymous user";
+
+        $headers = $this->createFeedbackHeaders($fromName, $fromEmail, $toMail, $sender);
         $params = empty($envelopeSender) ? '' : "-f $envelopeSender";
 
         // adding some information about the user for debugging purposes.
         $message = $message . "<br /><br /> Debugging information:"
             . "<br />Timestamp: " . date(DATE_RFC2822)
             . "<br />User agent: " . $request->getServerConstant('HTTP_USER_AGENT')
-            . "<br />IP address: " . $request->getServerConstant('REMOTE_ADDR')
             . "<br />Referer: " . $request->getServerConstant('HTTP_REFERER');
 
         try {
@@ -396,7 +406,7 @@ class WebController extends Controller
                 'search_results' => $searchResults,
                 'rest' => $parameters->getOffset()>0,
                 'global_search' => true,
-                'term' => $request->getQueryParam('q'),
+                'term' => $request->getQueryParamRaw('q'),
                 'lang_list' => $langList,
                 'vocabs' => str_replace(' ', '+', $vocabs),
                 'vocab_list' => $vocabList,
@@ -464,7 +474,7 @@ class WebController extends Controller
                 'limit_scheme' =>  $request->getQueryParam('scheme') ? explode('+', $request->getQueryParam('scheme')) : null,
                 'group_index' => $vocab->listConceptGroups($request->getContentLang()),
                 'parameters' => $parameters,
-                'term' => $request->getQueryParam('q'),
+                'term' => $request->getQueryParamRaw('q'),
                 'types' => $vocabTypes,
                 'explicit_langcodes' => $langcodes,
                 'request' => $request,
@@ -492,11 +502,15 @@ class WebController extends Controller
 
         $allAtOnce = $vocab->getConfig()->getAlphabeticalFull();
         if (!$allAtOnce) {
-            $searchResults = $vocab->searchConceptsAlphabetical($request->getLetter(), $count, $offset, $contentLang);
             $letters = $vocab->getAlphabet($contentLang);
+            $letter = $request->getLetter();
+            if ($letter === '') {
+                $letter = $letters[0];
+            }
+            $searchResults = $vocab->searchConceptsAlphabetical($letter, $count, $offset, $contentLang);
         } else {
-            $searchResults = $vocab->searchConceptsAlphabetical('*', null, null, $contentLang);
             $letters = null;
+            $searchResults = $vocab->searchConceptsAlphabetical('*', null, null, $contentLang);
         }
 
         $request->setContentLang($contentLang);
