@@ -40,6 +40,22 @@ class ConceptPropertyValue extends VocabularyDataObject
         if ($this->clang) {
             $lang = $this->clang;
         }
+        if ($this->vocab->getConfig()->getLanguageOrder($lang)) {
+            foreach ($this->vocab->getConfig()->getLanguageOrder($lang) as $fallback) {
+                if ($this->resource->label($fallback) !== null) {
+                    return $this->resource->label($fallback);
+                }
+                // We need to check all the labels in case one of them matches a subtag of the current language
+                if ($this->resource->allLiterals('skos:prefLabel')) {
+                    foreach($this->resource->allLiterals('skos:prefLabel') as $label) {
+                        // the label lang code is a subtag of the UI lang eg. en-GB - create a new literal with the main language
+                        if ($label !== null && strpos($label->getLang(), $fallback . '-') === 0) {
+                            return EasyRdf\Literal::create($label, $fallback);
+                        }
+                    }
+                }
+            }
+        }
 
         if ($this->resource->label($lang) !== null) { // current language
             return $this->resource->label($lang);
@@ -52,6 +68,7 @@ class ConceptPropertyValue extends VocabularyDataObject
         } elseif ($this->resource->getLiteral('rdf:value') !== null) { // any language
             return $this->resource->getLiteral('rdf:value');
         }
+        // uri if no label is found
         $label = $this->resource->shorten() ? $this->resource->shorten() : $this->getUri();
         return $label;
     }
@@ -116,15 +133,15 @@ class ConceptPropertyValue extends VocabularyDataObject
     public function isReified() {
         return (!$this->resource->label() && $this->resource->getLiteral('rdf:value'));
     }
-    
+
     public function getReifiedPropertyValues() {
         $ret = array();
         $props = $this->resource->propertyUris();
         foreach($props as $prop) {
             $prop = (EasyRdf\RdfNamespace::shorten($prop) !== null) ? EasyRdf\RdfNamespace::shorten($prop) : $prop;
             foreach ($this->resource->allLiterals($prop) as $val) {
-                if ($prop !== 'rdf:value' && $this->resource->get($prop)) { // shown elsewhere
-                    $ret[gettext($prop)] = new ConceptPropertyValueLiteral($this->resource->get($prop), $prop);
+                if ($prop !== 'rdf:value') { // shown elsewhere
+                    $ret[gettext($prop)] = new ConceptPropertyValueLiteral($this->model, $this->vocab, $this->resource, $val, $prop);
                 }
             }
             foreach ($this->resource->allResources($prop) as $val) {
