@@ -6,12 +6,12 @@ class PluginRegister {
     public function __construct($requestedPlugins=array()) {
         $this->requestedPlugins = $requestedPlugins;
     }
-    
+
     /**
      * Returns the plugin configurations found from plugin folders inside the plugins folder
      * @return array
      */
-    protected function getPlugins() 
+    protected function getPlugins()
     {
         $plugins = array();
         $pluginconfs = glob('plugins/*/plugin.json');
@@ -25,32 +25,38 @@ class PluginRegister {
     }
 
     /**
-     * Returns the plugin configurations found from plugin folders 
+     * Returns the plugin configurations found from plugin folders
      * inside the plugins folder filtered by filetype.
      * @param string $type filetype e.g. 'css', 'js' or 'template'
+     * @param boolean $raw interpret $type values as raw text instead of files
      * @return array
      */
-    private function filterPlugins($type) {
+    private function filterPlugins($type, $raw=false) {
         $plugins = $this->getPlugins();
         $ret = array();
         if (!empty($plugins)) {
             foreach ($plugins as $name => $files) {
                 if (isset($files[$type])) {
                     $ret[$name] = array();
-                    foreach ($files[$type] as $file) {
-                        array_push($ret[$name], 'plugins/' . $name . '/' . $file);
+                    if ($raw) {
+                        $ret[$name] = $files[$type];
+                    }
+                    else {
+                        foreach ($files[$type] as $file) {
+                            array_push($ret[$name], 'plugins/' . $name . '/' . $file);
+                        }
                     }
                 }
-            } 
+            }
         }
         return $ret;
     }
 
     /**
-     * Returns the plugin configurations found from plugin folders 
+     * Returns the plugin configurations found from plugin folders
      * inside the plugins folder filtered by plugin name (the folder name).
      * @param string $type filetype e.g. 'css', 'js' or 'template'
-     * @param array $names the plugin name strings (foldernames) in an array 
+     * @param array $names the plugin name strings (foldernames) in an array
      * @return array
      */
     private function filterPluginsByName($type, $names) {
@@ -65,7 +71,7 @@ class PluginRegister {
 
     /**
      * Returns an array of javascript filepaths
-     * @param array $names the plugin name strings (foldernames) in an array 
+     * @param array $names the plugin name strings (foldernames) in an array
      * @return array
      */
     public function getPluginsJS($names=null) {
@@ -78,7 +84,7 @@ class PluginRegister {
 
     /**
      * Returns an array of css filepaths
-     * @param array $names the plugin name strings (foldernames) in an array 
+     * @param array $names the plugin name strings (foldernames) in an array
      * @return array
      */
     public function getPluginsCSS($names=null) {
@@ -91,7 +97,7 @@ class PluginRegister {
 
     /**
      * Returns an array of template filepaths
-     * @param array $names the plugin name strings (foldernames) in an array 
+     * @param array $names the plugin name strings (foldernames) in an array
      * @return array
      */
     public function getPluginsTemplates($names=null) {
@@ -104,7 +110,7 @@ class PluginRegister {
 
     /**
      * Returns an array of template files contents as strings
-     * @param array $names the plugin name strings (foldernames) in an array 
+     * @param array $names the plugin name strings (foldernames) in an array
      * @return array
      */
     public function getTemplates($names=null) {
@@ -124,7 +130,7 @@ class PluginRegister {
     }
 
     /**
-     * Returns an array of javascript function names to call when loading a new concept 
+     * Returns an array of javascript function names to call when loading pages
      * @return array
      */
     public function getCallbacks() {
@@ -138,4 +144,25 @@ class PluginRegister {
         }
         return $ret;
     }
+ 
+    /**
+     * Returns an array that is flattened from its possibly multidimensional form
+     * copied from https://stackoverflow.com/a/1320156
+     * @param mixed[] $array Flattens this input array
+     * @return array Flattened input
+     */
+    protected function flatten($array) {
+        $return = array();
+        array_walk_recursive($array, function($a) use (&$return) { $return[] = $a; });
+        return $return;
+    }
+
+    /**
+     * Returns a flattened array containing the external properties we are interested in saving
+     * @return string[]
+     */
+    public function getExtProperties() {
+        return array_unique($this->flatten($this->filterPlugins('ext-properties', true)));
+    }
 }
+
